Added static_cast at dataset.cc IndexDataSet getpattern

+++ packages/basics/dataset/c_src/dataset.cc
@@ -463,7 +463,7 @@ int IndexDataSet<T>::getPattern(int index, T *pat) {
   int pos = 0;
   indices->getPattern(index,patternindices);
   for (int i=0; i < numdiccionarios; i++) {
-    int idx = patternindices[i]-firstindex;
+    int idx = static_cast<int>(patternindices[i])-firstindex;
     assert("Incorrect index at IndexDataSet" && idx >= 0);
     pos += diccionarios[i]->getPattern((int)idx,pat+pos);
   }
